nbd: fix might_sleep warning on socket shutdown
spinlocked ranges should be small and not contain calls into huge
subfunctions. Fix this and just get the pointer to the socket instead of
doing everything with spinlock held.

Use the system work queue to actually trigger the shutdown of socket.
This solves the issue when kernel_sendmsg is currently blocked while a
timeout occurs.

Reported-by: Mikulas Patocka <mikulas@twibright.com>
Signed-off-by: Pranay Kr Srivastava <pranjas@gmail.com>
Signed-off-by: Markus Pargmann <mpa.kernel@devmp.org>