nbd: make nbd device wait for its users
When a timeout occurs or a recv fails, then instead of abruptly killing
nbd block device wait for its users to finish.

This is more required when filesystem(s) like ext2 or ext3 don't expect
their buffer heads to disappear while the filesystem is mounted.

Each open is now refcounted with the device being released for re-use
only when there are no "other users".

A timedout or a disconnected device, if in use, can't be used until it
has been resetted. The reset happens when all tasks having this bdev
open closes this bdev.

Behavioral Change:

1)	NBD_DO_IT will not wait for the device to be reset. Hence
	the nbd-client "may exit" while some other process is using
	this device without actually doing the reset on this device
	, hence thus making it unusable until all such user space
	processes have stopped using this device.

2)	There's a window where the nbd-client will not be able to
	change / issue ioctls to the nbd device. This is when there's
	been a disconnect issued or a timeout has occurred, however
	there are "other" user processes which currently have this
	nbd device opened.

Signed-off-by: Pranay Kr Srivastava <pranjas@gmail.com>
Signed-off-by: Markus Pargmann <mpa.kernel@devmp.org>